🐛 avoid caching results
when only the last category is cycled, the browser caches the result (contents of alert box) and this works even when the network is cut off which shouldn't happen. However, when the 2nd category is also cycled along with 3rd, actual status is then fetched. So, from now on, fix the first category and cycle the next two.